feat: 프로젝트 레이어 구조 조정, mybatis, jpa 환경구축
