+++ core/core-api/src/main/java/io/dodn/springboot/common/api/HealthController.java
@@ -1,0 +1,18 @@
+package io.dodn.springboot.common.api;
+
+import io.dodn.springboot.common.swagger.HealthDocs;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HealthController implements HealthDocs {
+
+    @GetMapping("/health")
+    public ResponseEntity<Void> health() {
+
+        return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+}
