+++ core/core-api/src/main/java/io/dodn/springboot/common/support/ApiControllerAdvice.java
@@ -1,0 +1,34 @@
+package io.dodn.springboot.common.support;
+
+import io.dodn.springboot.common.support.error.CoreException;
+import io.dodn.springboot.common.support.error.ErrorType;
+import io.dodn.springboot.common.support.response.ApiResponse;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ApiControllerAdvice {
+
+    private final Logger log = LoggerFactory.getLogger(getClass());
+
+    @ExceptionHandler(CoreException.class)
+    public ResponseEntity<ApiResponse<?>> handleCoreException(CoreException e) {
+        switch (e.getErrorType().getLogLevel()) {
+            case ERROR -> log.error("CoreException : {}", e.getMessage(), e);
+            case WARN -> log.warn("CoreException : {}", e.getMessage(), e);
+            default -> log.info("CoreException : {}", e.getMessage(), e);
+        }
+        return new ResponseEntity<>(ApiResponse.error(e.getErrorType(), e.getData()), e.getErrorType().getStatus());
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ApiResponse<?>> handleException(Exception e) {
+        log.error("Exception : {}", e.getMessage(), e);
+        return new ResponseEntity<>(ApiResponse.error(ErrorType.DEFAULT_ERROR), ErrorType.DEFAULT_ERROR.getStatus());
+    }
+
+}
