+++ core/core-api/src/main/java/io/dodn/springboot/common/support/error/CoreException.java
@@ -1,0 +1,29 @@
+package io.dodn.springboot.common.support.error;
+
+public class CoreException extends RuntimeException {
+
+    private final ErrorType errorType;
+
+    private final Object data;
+
+    public CoreException(ErrorType errorType) {
+        super(errorType.getMessage());
+        this.errorType = errorType;
+        this.data = null;
+    }
+
+    public CoreException(ErrorType errorType, Object data) {
+        super(errorType.getMessage());
+        this.errorType = errorType;
+        this.data = data;
+    }
+
+    public ErrorType getErrorType() {
+        return errorType;
+    }
+
+    public Object getData() {
+        return data;
+    }
+
+}
