+++ core/core-api/src/main/java/io/dodn/springboot/common/support/error/ErrorCode.java
@@ -1,0 +1,7 @@
+package io.dodn.springboot.common.support.error;
+
+public enum ErrorCode {
+
+    E500
+
+}
