+++ core/core-api/src/main/java/io/dodn/springboot/common/support/response/ResultType.java
@@ -1,0 +1,7 @@
+package io.dodn.springboot.common.support.response;
+
+public enum ResultType {
+
+    SUCCESS, ERROR
+
+}
