--- core/core-api/src/main/java/io/dodn/springboot/core/api/controller/HealthController.java
@@ -1,16 +1,0 @@
-package io.dodn.springboot.core.api.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthController implements HealthDocs{
-
-    @GetMapping("/health")
-    public ResponseEntity<Void> health() {
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-}
