--- core/core-api/src/main/java/io/dodn/springboot/core/api/controller/v1/request/ExampleRequestDto.java
@@ -1,9 +1,0 @@
-package io.dodn.springboot.core.api.controller.v1.request;
-
-import io.dodn.springboot.core.domain.ExampleData;
-
-public record ExampleRequestDto(String data) {
-    public ExampleData toExampleData() {
-        return new ExampleData(data, data);
-    }
-}
