--- core/core-api/src/main/java/io/dodn/springboot/core/api/controller/v1/response/ExampleResponseDto.java
@@ -1,4 +1,0 @@
-package io.dodn.springboot.core.api.controller.v1.response;
-
-public record ExampleResponseDto(String result) {
-}
