+++ core/core-api/src/main/java/io/dodn/springboot/core/controller/request/ExampleRequestDto.java
@@ -1,0 +1,9 @@
+package io.dodn.springboot.core.controller.request;
+
+import io.dodn.springboot.core.domain.ExampleData;
+
+public record ExampleRequestDto(String data) {
+    public ExampleData toExampleData() {
+        return new ExampleData(data, data);
+    }
+}
