+++ core/core-api/src/main/java/io/dodn/springboot/core/controller/response/ExampleResponseDto.java
@@ -1,0 +1,4 @@
+package io.dodn.springboot.core.controller.response;
+
+public record ExampleResponseDto(String result) {
+}
