--- core/core-api/src/main/java/io/dodn/springboot/core/support/error/CoreException.java
@@ -1,29 +1,0 @@
-package io.dodn.springboot.core.support.error;
-
-public class CoreException extends RuntimeException {
-
-    private final ErrorType errorType;
-
-    private final Object data;
-
-    public CoreException(ErrorType errorType) {
-        super(errorType.getMessage());
-        this.errorType = errorType;
-        this.data = null;
-    }
-
-    public CoreException(ErrorType errorType, Object data) {
-        super(errorType.getMessage());
-        this.errorType = errorType;
-        this.data = data;
-    }
-
-    public ErrorType getErrorType() {
-        return errorType;
-    }
-
-    public Object getData() {
-        return data;
-    }
-
-}
