--- core/core-api/src/main/java/io/dodn/springboot/core/support/error/ErrorCode.java
@@ -1,7 +1,0 @@
-package io.dodn.springboot.core.support.error;
-
-public enum ErrorCode {
-
-    E500
-
-}
