--- core/core-api/src/main/java/io/dodn/springboot/core/support/error/ErrorMessage.java
@@ -1,35 +1,0 @@
-package io.dodn.springboot.core.support.error;
-
-public class ErrorMessage {
-
-    private final String code;
-
-    private final String message;
-
-    private final Object data;
-
-    public ErrorMessage(ErrorType errorType) {
-        this.code = errorType.getCode().name();
-        this.message = errorType.getMessage();
-        this.data = null;
-    }
-
-    public ErrorMessage(ErrorType errorType, Object data) {
-        this.code = errorType.getCode().name();
-        this.message = errorType.getMessage();
-        this.data = data;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public Object getData() {
-        return data;
-    }
-
-}
