--- core/core-api/src/main/java/io/dodn/springboot/core/support/response/ResultType.java
@@ -1,7 +1,0 @@
-package io.dodn.springboot.core.support.response;
-
-public enum ResultType {
-
-    SUCCESS, ERROR
-
-}
