--- storage/db-core/src/main/java/io/dodn/springboot/storage/db/core/config/CoreDataSourceConfig.java
@@ -1,25 +1,0 @@
-package io.dodn.springboot.storage.db.core.config;
-
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-class CoreDataSourceConfig {
-
-    @Bean
-    @ConfigurationProperties(prefix = "storage.datasource.core")
-    public HikariConfig coreHikariConfig() {
-        return new HikariConfig();
-    }
-
-    @Bean
-    public HikariDataSource coreDataSource(@Qualifier("coreHikariConfig") HikariConfig config) {
-        return new HikariDataSource(config);
-    }
-
-}
