+++ storage/db-core/src/main/java/io/dodn/springboot/storage/db/core/repository/ExampleJpaRepository.java
@@ -1,0 +1,8 @@
+package io.dodn.springboot.storage.db.core.repository;
+
+import io.dodn.springboot.storage.db.core.entity.ExampleEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ExampleJpaRepository extends JpaRepository<ExampleEntity, Long> {
+
+}
