+++ storage/db-core/src/main/java/io/dodn/springboot/storage/db/core/repository/ExampleMybatisRepository.java
@@ -1,0 +1,8 @@
+package io.dodn.springboot.storage.db.core.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface ExampleMybatisRepository {
+    int updateExampleColumn(Long id, String exampleColumn);
+}
