+++ storage/db-core/src/main/resources/mappers/ExampleMapper.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="io.dodn.springboot.storage.db.core.repository.ExampleMybatisRepository">
+
+    <update id="updateExampleColumn">
+        UPDATE example_table
+        SET example_column = #{exampleColumn}
+        WHERE id = #{id}
+    </update>
+</mapper>
